code/requirements.txt
jupyter
gradio
Pillow
facenet-pytorch

opencv-python
grad-cam
transformers
streamlit
mediapipe
opencv-python-headless
moviepy
Transformers 
Datasets
Tokenizers
librosa
pandas
audio-recorder-streamlit
